refactor: make the relational write path generic over sqlx::Database (phase 2)
The upsert/patch/delete/insert/update/row_version `*_in_tx` functions were
byte-identical between postgres_repo and sqlite_repo modulo the `Postgres`/
`Sqlite` type parameter — `QueryBuilder<DB>` already renders each dialect's
placeholder style. Hoist them into `sqlx_repo::read_model` as free functions
over `DB: SqlxReadModelBackend`.

Only the genuinely dialect-specific bits stay per-backend, behind the one
`SqlxReadModelBackend` trait (two impls, one line per method): value binding
(`Bool`/typed-`NULL` — Postgres binds native `bool` + `::jsonb`/`::timestamptz`
casts; SQLite stores bools as `i64` and collapses integer/bool NULLs), the
`rows_affected` accessor (sqlx exposes it only as an inherent method per
backend), and the backend/storage label strings. The unavoidable sqlx
executor/encode/decode bounds are stated explicitly on each write function —
no abstraction layer, no wrapper types.

Net ~263 lines removed (two ~470-line copies replaced by one generic copy plus
two small trait impls). No public API change; behavior verified identical on
real Postgres and SQLite.
