+++ src/postgres_repo/mod.rs
@@ -24,12 +24,10 @@ use crate::outbox_worker::{
     OutboxPublishFailureAction,
 };
 use crate::read_model::{
-    column_name_for, key_fingerprint, validate_key, validate_row_values, ColumnDef, ColumnType,
-    DeleteRowMutation, PatchMode, PatchRowMutation, ReadModelAdapterCapabilities,
+    column_name_for, validate_key, ColumnDef, ColumnType, ReadModelAdapterCapabilities,
     ReadModelCommitOutcome, ReadModelError, ReadModelIncludeRows, ReadModelLoadGraph,
-    ReadModelLoadRequest, ReadModelMutation, ReadModelQueryCapabilities, ReadModelSchema,
-    ReadModelWritePlan, RelationshipKind, RowKey, RowMutation, RowValue, RowValues, RowWriteMode,
-    Versioned,
+    ReadModelLoadRequest, ReadModelQueryCapabilities, ReadModelSchema, ReadModelWritePlan,
+    RelationshipKind, RowKey, RowValue, RowValues, Versioned,
 };
 use crate::repository::{
     reject_duplicate_outbox_messages, reject_duplicate_streams,
@@ -40,11 +38,11 @@ use crate::repository::{
 };
 use crate::snapshot::SnapshotRecord;
 use crate::sqlx_repo::read_model::{
-    belongs_to_target_column, column_by_name, empty_string_as_none, initial_row_version,
-    next_row_version, patch_values_preserving_key, quote_identifier, remember_read_model_schemas,
-    resolve_registered_read_model_schemas, row_concurrency_conflict, row_values_from_key_and_patch,
-    row_write_values, sql_read_model_capabilities, validate_row_expected_version,
-    validate_sql_write_plan, validate_values_match_key, version_column, IncludeSpec,
+    apply_read_model_write_plan_in_tx, begin_read_model_tx, belongs_to_target_column,
+    column_by_name, commit_read_model_tx, empty_string_as_none, push_key_predicates,
+    quote_identifier, remember_read_model_schemas, resolve_registered_read_model_schemas,
+    sql_read_model_capabilities, validate_sql_write_plan, version_column, IncludeSpec,
+    SqlxReadModelBackend,
 };
 use crate::sqlx_repo::{
     self, audited_table_schema_sql, deserialize_event_metadata, is_postgres_unique_violation,
@@ -946,381 +944,84 @@ async fn insert_inbox_receipt_in_tx(
     }
 }
 
-async fn begin_read_model_tx(pool: &PgPool) -> Result<Transaction<'_, Postgres>, ReadModelError> {
-    pool.begin()
-        .await
-        .map_err(|err| read_model_storage_error("begin transaction", err))
-}
+impl crate::sqlx_repo::read_model::SqlxReadModelBackend for Postgres {
+    const BACKEND: &'static str = POSTGRES_BACKEND;
+    const INTEGER_STORAGE: &'static str = BIGINT_STORAGE;
 
-async fn commit_read_model_tx(tx: Transaction<'_, Postgres>) -> Result<(), ReadModelError> {
-    tx.commit()
-        .await
-        .map_err(|err| read_model_storage_error("commit transaction", err))
-}
-
-async fn apply_read_model_write_plan_in_tx(
-    tx: &mut Transaction<'_, Postgres>,
-    plan: ReadModelWritePlan,
-) -> Result<ReadModelCommitOutcome, ReadModelError> {
-    validate_sql_write_plan(&plan)?;
-
-    for mutation in plan.mutations {
-        match mutation {
-            ReadModelMutation::UpsertRow(mutation) => {
-                upsert_relational_row_in_tx(tx, mutation).await?;
+    fn push_row_value_bind(
+        builder: &mut QueryBuilder<Postgres>,
+        value: RowValue,
+        column: &ColumnDef,
+    ) -> Result<(), ReadModelError> {
+        match value {
+            RowValue::Null => Self::push_null_bind(builder, column)?,
+            RowValue::Bool(value) => {
+                builder.push_bind(value);
             }
-            ReadModelMutation::PatchRow(mutation) => {
-                patch_relational_row_in_tx(tx, mutation).await?;
+            RowValue::I64(value) => {
+                builder.push_bind(value);
             }
-            ReadModelMutation::DeleteRow(mutation) => {
-                delete_relational_row_in_tx(tx, mutation).await?;
+            RowValue::U64(value) => {
+                builder.push_bind(sqlx_read_model_i64_from_u64(
+                    POSTGRES_BACKEND,
+                    value,
+                    &column.column_name,
+                    BIGINT_STORAGE,
+                )?);
             }
-        }
-    }
-
-    Ok(ReadModelCommitOutcome::applied())
-}
-
-async fn upsert_relational_row_in_tx(
-    tx: &mut Transaction<'_, Postgres>,
-    mutation: RowMutation,
-) -> Result<(), ReadModelError> {
-    validate_key(&mutation.schema, &mutation.key)?;
-    validate_row_values(&mutation.schema, &mutation.values, true)?;
-    validate_values_match_key(&mutation.schema, &mutation.key, &mutation.values)?;
-
-    let current_version = row_version_in_tx(tx, &mutation.schema, &mutation.key).await?;
-    validate_row_expected_version(
-        &mutation.schema,
-        &mutation.key,
-        &mutation.expected_version,
-        current_version,
-    )?;
-    if matches!(mutation.mode, RowWriteMode::Insert) && current_version.is_some() {
-        return Err(row_concurrency_conflict(
-            &mutation.schema,
-            &mutation.key,
-            0,
-            current_version.unwrap_or_default(),
-        ));
-    }
-
-    match current_version {
-        Some(expected_version) => {
-            let new_version = next_row_version(&mutation.schema, &mutation.key, current_version)?;
-            let rows_affected = update_relational_row_values_in_tx(
-                tx,
-                &mutation.schema,
-                &mutation.key,
-                &mutation.values,
-                expected_version,
-                new_version,
-            )
-            .await?;
-            if rows_affected == 0 {
-                let actual = row_version_in_tx(tx, &mutation.schema, &mutation.key)
-                    .await?
-                    .unwrap_or(expected_version);
-                return Err(row_concurrency_conflict(
-                    &mutation.schema,
-                    &mutation.key,
-                    expected_version,
-                    actual,
-                ));
+            RowValue::F64(value) => {
+                builder.push_bind(value);
+            }
+            RowValue::String(value) => {
+                builder.push_bind(value);
+            }
+            RowValue::Bytes(value) => {
+                builder.push_bind(value);
+            }
+            RowValue::Json(value) => {
+                let payload = serde_json::to_string(&value)
+                    .map_err(|err| ReadModelError::Serde(err.to_string()))?;
+                builder.push_bind(payload);
             }
         }
-        None => {
-            insert_relational_row_in_tx(
-                tx,
-                &mutation.schema,
-                &mutation.values,
-                initial_row_version(),
-            )
-            .await?;
-        }
+        push_postgres_type_cast(builder, column);
+        Ok(())
     }
 
-    Ok(())
-}
-
-async fn patch_relational_row_in_tx(
-    tx: &mut Transaction<'_, Postgres>,
-    mutation: PatchRowMutation,
-) -> Result<(), ReadModelError> {
-    validate_key(&mutation.schema, &mutation.key)?;
-
-    let current_version = row_version_in_tx(tx, &mutation.schema, &mutation.key).await?;
-    validate_row_expected_version(
-        &mutation.schema,
-        &mutation.key,
-        &mutation.expected_version,
-        current_version,
-    )?;
-
-    match current_version {
-        Some(expected_version) => {
-            let patch_values =
-                patch_values_preserving_key(&mutation.schema, &mutation.key, mutation.patch)?;
-            let new_version = next_row_version(&mutation.schema, &mutation.key, current_version)?;
-            let rows_affected = update_relational_patch_in_tx(
-                tx,
-                &mutation.schema,
-                &mutation.key,
-                patch_values,
-                expected_version,
-                new_version,
-            )
-            .await?;
-            if rows_affected == 0 {
-                let actual = row_version_in_tx(tx, &mutation.schema, &mutation.key)
-                    .await?
-                    .unwrap_or(expected_version);
-                return Err(row_concurrency_conflict(
-                    &mutation.schema,
-                    &mutation.key,
-                    expected_version,
-                    actual,
-                ));
+    fn push_null_bind(
+        builder: &mut QueryBuilder<Postgres>,
+        column: &ColumnDef,
+    ) -> Result<(), ReadModelError> {
+        match &column.column_type {
+            ColumnType::Text | ColumnType::Json | ColumnType::Timestamp => {
+                builder.push_bind(Option::<String>::None);
+            }
+            ColumnType::Boolean => {
+                builder.push_bind(Option::<bool>::None);
+            }
+            ColumnType::Integer | ColumnType::UnsignedInteger => {
+                builder.push_bind(Option::<i64>::None);
+            }
+            ColumnType::Float => {
+                builder.push_bind(Option::<f64>::None);
+            }
+            ColumnType::Bytes => {
+                builder.push_bind(Option::<Vec<u8>>::None);
+            }
+            ColumnType::Unsupported(type_name) => {
+                return Err(ReadModelError::Metadata(format!(
+                    "read model `{}` column `{}` has unsupported type `{}`",
+                    column.field_name, column.column_name, type_name
+                )));
             }
         }
-        None if matches!(mutation.mode, PatchMode::InsertMissing) => {
-            let values =
-                row_values_from_key_and_patch(&mutation.schema, &mutation.key, mutation.patch)?;
-            insert_relational_row_in_tx(tx, &mutation.schema, &values, initial_row_version())
-                .await?;
-        }
-        None => {
-            return Err(ReadModelError::NotFound {
-                collection: mutation.schema.table_name,
-                id: key_fingerprint(&mutation.key),
-            });
-        }
+        Ok(())
     }
 
-    Ok(())
+    fn rows_affected(result: &sqlx::postgres::PgQueryResult) -> u64 {
+        result.rows_affected()
+    }
 }
-
-async fn delete_relational_row_in_tx(
-    tx: &mut Transaction<'_, Postgres>,
-    mutation: DeleteRowMutation,
-) -> Result<(), ReadModelError> {
-    validate_key(&mutation.schema, &mutation.key)?;
-    let current_version = row_version_in_tx(tx, &mutation.schema, &mutation.key).await?;
-    validate_row_expected_version(
-        &mutation.schema,
-        &mutation.key,
-        &mutation.expected_version,
-        current_version,
-    )?;
-
-    let rows_affected = delete_relational_row_where_current_in_tx(
-        tx,
-        &mutation.schema,
-        &mutation.key,
-        current_version,
-    )
-    .await?;
-    if rows_affected == 0 {
-        if let Some(expected_version) = current_version {
-            let actual = row_version_in_tx(tx, &mutation.schema, &mutation.key)
-                .await?
-                .unwrap_or(expected_version);
-            return Err(row_concurrency_conflict(
-                &mutation.schema,
-                &mutation.key,
-                expected_version,
-                actual,
-            ));
-        }
-    }
-
-    Ok(())
-}
-
-async fn row_version_in_tx(
-    tx: &mut Transaction<'_, Postgres>,
-    schema: &ReadModelSchema,
-    key: &RowKey,
-) -> Result<Option<u64>, ReadModelError> {
-    let version_column = version_column(schema)?;
-    let mut builder = QueryBuilder::<Postgres>::new("SELECT ");
-    builder.push(quote_identifier(version_column));
-    builder.push(" FROM ");
-    builder.push(quote_identifier(&schema.table_name));
-    push_postgres_key_predicates(&mut builder, schema, key)?;
-
-    let row = builder
-        .build()
-        .fetch_optional(&mut **tx)
-        .await
-        .map_err(|err| read_model_storage_error("load relational row version", err))?;
-
-    row.map(|row| {
-        sqlx_read_model_u64_from_i64(
-            POSTGRES_BACKEND,
-            row.try_get::<i64, _>(version_column)
-                .map_err(|err| read_model_storage_error("decode relational row version", err))?,
-            version_column,
-        )
-    })
-    .transpose()
-}
-
-async fn insert_relational_row_in_tx(
-    tx: &mut Transaction<'_, Postgres>,
-    schema: &ReadModelSchema,
-    values: &RowValues,
-    version: u64,
-) -> Result<(), ReadModelError> {
-    let version_column = version_column(schema)?;
-    let write_values = row_write_values(schema, values)?;
-    let has_write_values = !write_values.is_empty();
-    let mut builder = QueryBuilder::<Postgres>::new("INSERT INTO ");
-    builder.push(quote_identifier(&schema.table_name));
-    builder.push(" (");
-    for (index, (column, _)) in write_values.iter().enumerate() {
-        if index > 0 {
-            builder.push(", ");
-        }
-        builder.push(quote_identifier(&column.column_name));
-    }
-    if has_write_values {
-        builder.push(", ");
-    }
-    builder.push(quote_identifier(version_column));
-    builder.push(") VALUES (");
-    for (index, (column, value)) in write_values.into_iter().enumerate() {
-        if index > 0 {
-            builder.push(", ");
-        }
-        push_postgres_row_value_bind(&mut builder, value, column)?;
-    }
-    if has_write_values {
-        builder.push(", ");
-    }
-    builder.push_bind(sqlx_read_model_i64_from_u64(
-        POSTGRES_BACKEND,
-        version,
-        version_column,
-        BIGINT_STORAGE,
-    )?);
-    builder.push(")");
-
-    builder
-        .build()
-        .execute(&mut **tx)
-        .await
-        .map_err(|err| read_model_storage_error("insert relational row", err))?;
-
-    Ok(())
-}
-
-async fn update_relational_row_values_in_tx(
-    tx: &mut Transaction<'_, Postgres>,
-    schema: &ReadModelSchema,
-    key: &RowKey,
-    values: &RowValues,
-    expected_version: u64,
-    version: u64,
-) -> Result<u64, ReadModelError> {
-    let mut write_values = row_write_values(schema, values)?;
-    write_values.retain(|(column, _)| !column.primary_key);
-    update_relational_columns_in_tx(tx, schema, key, write_values, expected_version, version).await
-}
-
-async fn update_relational_patch_in_tx(
-    tx: &mut Transaction<'_, Postgres>,
-    schema: &ReadModelSchema,
-    key: &RowKey,
-    write_values: Vec<(&ColumnDef, RowValue)>,
-    expected_version: u64,
-    version: u64,
-) -> Result<u64, ReadModelError> {
-    update_relational_columns_in_tx(tx, schema, key, write_values, expected_version, version).await
-}
-
-async fn update_relational_columns_in_tx(
-    tx: &mut Transaction<'_, Postgres>,
-    schema: &ReadModelSchema,
-    key: &RowKey,
-    write_values: Vec<(&ColumnDef, RowValue)>,
-    expected_version: u64,
-    version: u64,
-) -> Result<u64, ReadModelError> {
-    let version_column = version_column(schema)?;
-    let mut builder = QueryBuilder::<Postgres>::new("UPDATE ");
-    builder.push(quote_identifier(&schema.table_name));
-    builder.push(" SET ");
-    let mut wrote_set = false;
-    for (column, value) in write_values {
-        if wrote_set {
-            builder.push(", ");
-        }
-        builder.push(quote_identifier(&column.column_name));
-        builder.push(" = ");
-        push_postgres_row_value_bind(&mut builder, value, column)?;
-        wrote_set = true;
-    }
-    if wrote_set {
-        builder.push(", ");
-    }
-    builder.push(quote_identifier(version_column));
-    builder.push(" = ");
-    builder.push_bind(sqlx_read_model_i64_from_u64(
-        POSTGRES_BACKEND,
-        version,
-        version_column,
-        BIGINT_STORAGE,
-    )?);
-    push_postgres_key_predicates(&mut builder, schema, key)?;
-    builder.push(" AND ");
-    builder.push(quote_identifier(version_column));
-    builder.push(" = ");
-    builder.push_bind(sqlx_read_model_i64_from_u64(
-        POSTGRES_BACKEND,
-        expected_version,
-        "expected version",
-        BIGINT_STORAGE,
-    )?);
-
-    let result = builder
-        .build()
-        .execute(&mut **tx)
-        .await
-        .map_err(|err| read_model_storage_error("update relational row", err))?;
-    Ok(result.rows_affected())
-}
-
-async fn delete_relational_row_where_current_in_tx(
-    tx: &mut Transaction<'_, Postgres>,
-    schema: &ReadModelSchema,
-    key: &RowKey,
-    current_version: Option<u64>,
-) -> Result<u64, ReadModelError> {
-    let mut builder = QueryBuilder::<Postgres>::new("DELETE FROM ");
-    builder.push(quote_identifier(&schema.table_name));
-    push_postgres_key_predicates(&mut builder, schema, key)?;
-    if let Some(version) = current_version {
-        let version_column = version_column(schema)?;
-        builder.push(" AND ");
-        builder.push(quote_identifier(version_column));
-        builder.push(" = ");
-        builder.push_bind(sqlx_read_model_i64_from_u64(
-            POSTGRES_BACKEND,
-            version,
-            "expected version",
-            BIGINT_STORAGE,
-        )?);
-    }
-
-    let result = builder
-        .build()
-        .execute(&mut **tx)
-        .await
-        .map_err(|err| read_model_storage_error("delete relational row", err))?;
-    Ok(result.rows_affected())
-}
-
 async fn load_postgres_relational_row_by_key(
     pool: &PgPool,
     schema: &ReadModelSchema,
@@ -1328,7 +1029,7 @@ async fn load_postgres_relational_row_by_key(
 ) -> Result<Option<Versioned<RowValues>>, ReadModelError> {
     validate_key(schema, key)?;
     let mut builder = postgres_relational_row_select(schema)?;
-    push_postgres_key_predicates(&mut builder, schema, key)?;
+    push_key_predicates(&mut builder, schema, key)?;
     let row = builder
         .build()
         .fetch_optional(pool)
@@ -1435,7 +1136,7 @@ async fn load_postgres_rows_matching_column(
     builder.push(" WHERE ");
     builder.push(quote_identifier(column_name));
     builder.push(" = ");
-    push_postgres_row_value_bind(&mut builder, value.clone(), column)?;
+    <Postgres as SqlxReadModelBackend>::push_row_value_bind(&mut builder, value.clone(), column)?;
     push_postgres_order_by_primary_key(&mut builder, schema);
 
     let rows = builder
@@ -1570,100 +1271,6 @@ fn postgres_row_value(row: &PgRow, column: &ColumnDef) -> Result<RowValue, ReadM
             )));
         }
     })
-}
-
-fn push_postgres_key_predicates(
-    builder: &mut QueryBuilder<Postgres>,
-    schema: &ReadModelSchema,
-    key: &RowKey,
-) -> Result<(), ReadModelError> {
-    builder.push(" WHERE ");
-    for (index, column_name) in schema.primary_key.columns.iter().enumerate() {
-        if index > 0 {
-            builder.push(" AND ");
-        }
-        let column = column_by_name(schema, column_name)?;
-        let value = key.get(column_name).cloned().ok_or_else(|| {
-            ReadModelError::Metadata(format!(
-                "read model `{}` row key is missing primary-key column `{}`",
-                schema.model_name, column_name
-            ))
-        })?;
-        builder.push(quote_identifier(column_name));
-        builder.push(" = ");
-        push_postgres_row_value_bind(builder, value, column)?;
-    }
-    Ok(())
-}
-
-fn push_postgres_row_value_bind(
-    builder: &mut QueryBuilder<Postgres>,
-    value: RowValue,
-    column: &ColumnDef,
-) -> Result<(), ReadModelError> {
-    match value {
-        RowValue::Null => push_postgres_null_bind(builder, column)?,
-        RowValue::Bool(value) => {
-            builder.push_bind(value);
-        }
-        RowValue::I64(value) => {
-            builder.push_bind(value);
-        }
-        RowValue::U64(value) => {
-            builder.push_bind(sqlx_read_model_i64_from_u64(
-                POSTGRES_BACKEND,
-                value,
-                &column.column_name,
-                BIGINT_STORAGE,
-            )?);
-        }
-        RowValue::F64(value) => {
-            builder.push_bind(value);
-        }
-        RowValue::String(value) => {
-            builder.push_bind(value);
-        }
-        RowValue::Bytes(value) => {
-            builder.push_bind(value);
-        }
-        RowValue::Json(value) => {
-            let payload = serde_json::to_string(&value)
-                .map_err(|err| ReadModelError::Serde(err.to_string()))?;
-            builder.push_bind(payload);
-        }
-    }
-    push_postgres_type_cast(builder, column);
-    Ok(())
-}
-
-fn push_postgres_null_bind(
-    builder: &mut QueryBuilder<Postgres>,
-    column: &ColumnDef,
-) -> Result<(), ReadModelError> {
-    match &column.column_type {
-        ColumnType::Text | ColumnType::Json | ColumnType::Timestamp => {
-            builder.push_bind(Option::<String>::None);
-        }
-        ColumnType::Boolean => {
-            builder.push_bind(Option::<bool>::None);
-        }
-        ColumnType::Integer | ColumnType::UnsignedInteger => {
-            builder.push_bind(Option::<i64>::None);
-        }
-        ColumnType::Float => {
-            builder.push_bind(Option::<f64>::None);
-        }
-        ColumnType::Bytes => {
-            builder.push_bind(Option::<Vec<u8>>::None);
-        }
-        ColumnType::Unsupported(type_name) => {
-            return Err(ReadModelError::Metadata(format!(
-                "read model `{}` column `{}` has unsupported type `{}`",
-                column.field_name, column.column_name, type_name
-            )));
-        }
-    }
-    Ok(())
 }
 
 fn push_postgres_type_cast(builder: &mut QueryBuilder<Postgres>, column: &ColumnDef) {
